refactor: move auth components to (auth)/_components directory
- Move all auth components from app/components/auth/ to app/(auth)/_components/
- Update imports in sign-in and sign-up pages to use relative paths
- Better organization by keeping auth components within auth route group
- Maintains component functionality while improving project structure
