+++ apps/chat/src/app/(auth)/_components/sign-up-email-input.tsx
