+++ apps/chat/src/app/(auth)/sign-up/page.tsx
@@ -1,4 +1,4 @@
-import { SignUpForm } from "~/app/components/auth/sign-up-form";
+import { SignUpForm } from "../_components/sign-up-form";
 
 export default function SignUpPage() {
 	return <SignUpForm />;
